fix: Update scheduler and processes to use float for delta time

+++ src/Game/Game.cpp
@@ -44,7 +44,7 @@ void Game::run()
     SDL_Event event;
 
     // Create EnTT process scheduler
-    entt::scheduler<Uint64> scheduler;
+    entt::scheduler<float> scheduler;
 
     auto &updateProcess = scheduler.attach<UpdateProcess>(registry_);
     auto &renderProcess = scheduler.attach<RenderProcess>(registry_, renderer_.get());
